test(sam): move test
