+++ packages/faas-penguin-aws-lambda/app.js
@@ -1,0 +1,2 @@
+const { faasPenguinAwsLambda } = require("./lib/faas-penguin-aws-lambda");
+exports.lambdaHandler = faasPenguinAwsLambda;
