--- packages/faas-penguin-aws-lambda/hello-world/.npmignore
@@ -1,1 +1,0 @@
-tests/*
