--- packages/faas-penguin-aws-lambda/hello-world/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "hello_world",
-  "version": "1.0.0",
-  "description": "hello world sample for NodeJS",
-  "main": "app.js",
-  "repository": "https://github.com/awslabs/aws-sam-cli/tree/develop/samcli/local/init/templates/cookiecutter-aws-sam-hello-nodejs",
-  "author": "SAM CLI",
-  "license": "MIT",
-  "dependencies": {
-    "axios": "^0.21.1"
-  },
-  "scripts": {
-    "test": "mocha tests/unit"
-  },
-  "devDependencies": {
-    "chai": "^4.2.0",
-    "mocha": "^8.2.1"
-  }
-}
