+++ packages/faas-penguin-aws-lambda/lib/faas-penguin-aws-lambda.js
@@ -1,1 +1,30 @@
-export function faasPenguinAwsLambda() {}
+/**
+ *
+ * Event doc: https://docs.aws.amazon.com/apigateway/latest/developerguide/set-up-lambda-proxy-integrations.html#api-gateway-simple-proxy-for-lambda-input-format
+ * @param {Object} event - API Gateway Lambda Proxy Input Format
+ *
+ * Context doc: https://docs.aws.amazon.com/lambda/latest/dg/nodejs-prog-model-context.html
+ * @param {Object} context
+ *
+ * Return doc: https://docs.aws.amazon.com/apigateway/latest/developerguide/set-up-lambda-proxy-integrations.html
+ * @returns {Object} object - API Gateway Lambda Proxy Output Format
+ *
+ */
+exports.faasPenguinAwsLambda = async (event, context) => {
+  let response;
+  try {
+    // const ret = await axios(url);
+    response = {
+      statusCode: 200,
+      body: JSON.stringify({
+        message: "hello world from lib",
+        // location: ret.data.trim()
+      }),
+    };
+  } catch (err) {
+    console.log(err);
+    return err;
+  }
+
+  return response;
+};
