+++ packages/faas-penguin-aws-lambda/template.yaml
@@ -14,7 +14,7 @@ Resources:
   HelloWorldFunction:
     Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
     Properties:
-      CodeUri: hello-world/
+      CodeUri: ./
       Handler: app.lambdaHandler
       Runtime: nodejs14.x
       Events:
